requirements.txt
click==8.0.4
Flask==2.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.1.2
gunicorn==20.1.0
Pyrebase4==4.5.0
python-dotenv==0.21.0